[psale work] Added 'c' language switching
The user is now able to toggle the examples list from ASM ('s')~ default
~ to C ('c') by pressing Ctrl+S on the main form. The database does not
contain 'c' examples yet but that will be fixed next. Also, the 'fprin'
module got a structure (called 'FormularPrincipal') to bring it inline
with the other form entities and make the processing a little bit more
uniform.