[psale work] Added a first version of installer
Made with NSIS. It's self contained with bitmaps and icons but it lacks
the CD folder in 'construit' folder. I don't think that's a good ideea
to add to the repository. The commit also contains a windows-tools
folder (wtooluri) that has the 7za compressor. The Makefiles were
configured to compress psAle, to distribute it into it's designated
folder and to build the installer. For the installer, you have to have
NSIS installed since it uses makensis.exe. 'principal/sursele
mele/ale.h' was added for completeness but it doesn't contain working
code yet.